+++ src/snippet.js
@@ -2,21 +2,30 @@
  * @import { Issue, Article, User } from "@jetbrains/youtrack-scripting-api/entities" 
  * 
  * @typedef {{
- *  article: Article | null;                          // Article if snippet is in article
- *  issue: Issue | null;                              // Issue if snippet is in issue
- *  currentUser: User | null;                         // Current user
- *  refreshCount: number                              // Refresh count
- *  userInput?: string | number | boolean             // User input value
+ *  article?: Article | null;                             // Article if snippet is in article
+ *  issue?: Issue | null;                                 // Issue if snippet is in issue
+ *  currentUser?: User | null;                            // Current user
+ *  refreshCount: number                                  // Refresh count
+ *  userInput?: string | number | boolean                 // User input value
  * }} SnippetRuleContext
+ * 
+ * @typedef {(ctx: SnippetRuleContext) => string[]} EnumResolver
+ * 
+ * @typedef {SnippetRuleContext & { 
+ *  setUserInput: (value?: SnippetUserInput) => void      // Callback function to set user input
+ * }} SnippetGuardContext
+ * 
+ * @typedef {{
+ *  type: "string" | "number" | "boolean" | "text";       // Type of the user input
+ *  enum?: string[] | number[] | EnumResolver;            // Enum values for the user input
+ *  description: string;                                  // Description of the user input
+ *  required?: boolean;                                   // Whether the user input is required
+ * }} SnippetUserInput
  * 
  * @typedef {{
  *  name: string;                                         // Name of the snippet, should be unique in YouTrack
  *  title: string;                                        // Title of the snippet
- *  userInput?: {
- *    type: "string" | "number" | "boolean" | "text";     // Type of the user input
- *    enum?: string[] | number[];                         // Enum values for the user input
- *    description: string;                                // Description of the user input
- *  };
+ *  userInput?: SnippetUserInput;                         // User input for the snippet
  *  action: (ctx: SnippetRuleContext) => string;          // Action to be performed
  *  debug?: {
  *    guard?: (ctx: SnippetRuleContext) => boolean;       // Guard to run snippet in debug mode
@@ -25,18 +34,35 @@
  *    refreshCount?: number;                              // Refresh count to run snippet in debug mode
  *  }
  * }} SnippetRule
+ * 
+ * @typedef {{
+ *  title: string;                                        // Title of the snippet
+ *  command: string;                                      // Command to run the snippet
+ *  input?: {
+ *    type: "string" | "number" | "boolean";              // Type of the user input
+ *    description: string;                                // Description of the user input
+ *  };
+ *  action: (ctx: SnippetRuleContext) => string | void;   // Action to be performed
+ *  guard: (ctx: SnippetGuardContext) => boolean;         // Guard to run snippet
+ *  ruleType: "action";                                   // Rule type
+ *  target: "Issue" | "Article";                          // Target entity type
+ * }} ActionRule
  */
 
 class Snippet {
   /**
    * Embeds a snippet into the markdown content
    * @param {SnippetRule} rule
+   * @returns {ActionRule}
    */
   static forMarkdown(rule) {
+    if (!rule || typeof rule !== "object") {
+      throw new Error("Snippet rule must be an object with title, name and action properties");
+    }
     const { title, name, action, userInput, debug } = rule;
 
     if (!title || !name || !action) {
-      throw new Error("Snippet rule must have title, name and action");
+      throw new Error("Snippet rule must have title, name and action properties");
     }
 
     if (typeof action !== "function") {
@@ -55,10 +81,10 @@ class Snippet {
           throw new Error("Snippet rule userInput enum is only supported for string and number types");
         }
 
-        if (!Array.isArray(userInput.enum)) {
-          throw new Error("Snippet rule userInput enum must be an array");
+        if (!Array.isArray(userInput.enum) && typeof userInput.enum !== "function") {
+          throw new Error("Snippet rule userInput enum must be an array or a function");
         }
-        if (userInput.enum.some((v) => typeof v !== "string" && typeof v !== "number")) {
+        if (typeof userInput.enum !== "function" && userInput.enum.some((v) => typeof v !== "string" && typeof v !== "number")) {
           throw new Error("Snippet rule userInput enum must contain only strings or numbers");
         }
       }
@@ -67,15 +93,15 @@ class Snippet {
     return {
       title: `${debug ? "debug-" : ""}snippet:${title}`,
       command: `${debug ? "debug-" : ""}snippet:${name}`,
-      action: debug ? (/** @type {any} */ ctx) => {
+      action: debug ? (ctx) => {
         const result = action({ ...ctx, userInput: debug.userInput, refreshCount: debug.refreshCount ?? 0 });
-        if (debug.userInput === "Article") {
+        if (debug.userInput === "Article" && ctx.article) {
           ctx.article.content = result;
-        } else {
+        } else if (ctx.issue) {
           ctx.issue.description = result;
         }
       } : action,
-      guard: (/** @type {any} */ { setUserInput, ...rest }) => {
+      guard: ({ setUserInput, ...rest }) => {
         setUserInput?.(userInput);
 
         return debug?.guard ? debug.guard({ ...rest, userInput: debug.userInput }) : false;
@@ -98,7 +124,8 @@ exports.Snippet = Snippet;
 //   userInput: {
 //   	type: "string",
 //     enum: ["foo", "bar", "baz"],
-//     description: "Select an option"
+//     description: "Select an option",
+//     required: true
 //   },
 //   action: ({ issue, article, currentUser, userInput, refreshCount }) => {
 //     return [